content: extract popup manager scaffolding
