+++ internal/infrastructure/cef/content_injector.go
@@ -385,398 +385,6 @@ const redditDirectVideoJS = `(function(){
   }, 1000);
 })();`
 
-// clipboardSelectionFetchBridgeJS keeps the lightweight fetch bridge enabled
-// while the native renderer bridge remains disabled due to an OSR startup
-// regression. It mirrors explicit clipboard text from DOM selections,
-// safe input/textarea selections, and async Clipboard API writes, and also
-// reasserts browser focus when an editable element gains DOM focus.
-const clipboardSelectionFetchBridgeJS = `(function(){
-  var bridgeNonce = '__DUMBER_BRIDGE_NONCE__';
-
-  function encodeBody(body) {
-    return typeof btoa === 'function' ? btoa(unescape(encodeURIComponent(body))) : '';
-  }
-
-  function inputType(node) {
-    return node && node.type ? String(node.type).toLowerCase() : 'text';
-  }
-
-  function isTextEditableInputType(type) {
-    switch ((type || 'text').toLowerCase()) {
-    case 'button':
-    case 'submit':
-    case 'reset':
-    case 'checkbox':
-    case 'radio':
-    case 'range':
-    case 'color':
-    case 'file':
-    case 'image':
-    case 'hidden':
-      return false;
-    default:
-      return true;
-    }
-  }
-
-  function isSelectionReadableInputType(type) {
-    return isTextEditableInputType(type) && type !== 'password';
-  }
-
-  function getActiveElementSelection() {
-    var el = document.activeElement;
-    if (!el || el.nodeType !== 1 || typeof el.value !== 'string') return '';
-    var tag = el.tagName;
-    if (el.disabled || el.readOnly) return '';
-    if (tag === 'INPUT' && !isSelectionReadableInputType(inputType(el))) return '';
-    if (tag !== 'INPUT' && tag !== 'TEXTAREA') return '';
-    var start = typeof el.selectionStart === 'number' ? el.selectionStart : 0;
-    var end = typeof el.selectionEnd === 'number' ? el.selectionEnd : 0;
-    if (end > start) return el.value.slice(start, end);
-    return '';
-  }
-
-  function getSelectedText() {
-    var activeSelection = getActiveElementSelection();
-    if (activeSelection) return activeSelection;
-    var sel = window.getSelection();
-    return sel ? sel.toString() : '';
-  }
-
-  function isEditable(node) {
-    if (!node || node.nodeType !== 1) return false;
-    if (node.isContentEditable) return true;
-    var tag = node.tagName;
-    if (tag === 'TEXTAREA') return !node.disabled && !node.readOnly;
-    if (tag !== 'INPUT') return false;
-    if (node.disabled || node.readOnly) return false;
-    return isTextEditableInputType(inputType(node));
-  }
-
-  function sendToClipboard(text) {
-    if (!text) return;
-    var body = JSON.stringify({text: text});
-    var encoded = encodeBody(body);
-    if (!encoded) return;
-    fetch('dumb:///api/clipboard-set', {
-      method: 'POST',
-      headers: {
-        'X-Dumber-Body': encoded,
-        'X-Dumber-Bridge-Nonce': bridgeNonce
-      }
-    }).catch(function(){});
-  }
-
-  function sendFocusSync() {
-    fetch('dumb:///api/focus-sync', {
-      method: 'POST',
-      headers: {
-        'X-Dumber-Bridge-Action': 'focus-sync',
-        'X-Dumber-Bridge-Nonce': bridgeNonce
-      }
-    }).catch(function(){});
-  }
-
-  function mirrorClipboardItems(items) {
-    try {
-      if (!items || typeof items.length !== 'number') return;
-      Array.prototype.forEach.call(items, function(item) {
-        if (!item || !item.types || item.types.indexOf('text/plain') === -1 || typeof item.getType !== 'function') return;
-        Promise.resolve(item.getType('text/plain'))
-          .then(function(blob) {
-            if (!blob || typeof blob.text !== 'function') return '';
-            return blob.text();
-          })
-          .then(function(text) {
-            if (text) sendToClipboard(text);
-          })
-          .catch(function(){});
-      });
-    } catch (_) {}
-  }
-
-  document.addEventListener('copy', function(event) {
-    if (event && event.isTrusted === false) return;
-    var text = getSelectedText();
-    if (text) sendToClipboard(text);
-  });
-  document.addEventListener('cut', function(event) {
-    if (event && event.isTrusted === false) return;
-    var text = getSelectedText();
-    if (text) sendToClipboard(text);
-  });
-  document.addEventListener('focusin', function(event) {
-    if (event && event.isTrusted === false) return;
-    if (isEditable(event && event.target)) sendFocusSync();
-  }, true);
-
-  try {
-    var clipboardObj = navigator && navigator.clipboard ? navigator.clipboard : null;
-    var clipboardProto = (typeof Clipboard !== 'undefined' && Clipboard.prototype) ||
-      (clipboardObj && typeof Object.getPrototypeOf === 'function' ? Object.getPrototypeOf(clipboardObj) : null);
-
-    if (clipboardProto && typeof clipboardProto.writeText === 'function' && !clipboardProto.__dumberWriteTextPatched) {
-      clipboardProto.__dumberWriteTextPatched = true;
-      var originalProtoWriteText = clipboardProto.writeText;
-      var wrappedProtoWriteText = function(text) {
-        var normalized = text == null ? '' : String(text);
-        var result = originalProtoWriteText.call(this, normalized);
-        if (result && typeof result.then === 'function') {
-          return result.then(function(value) {
-            if (normalized) sendToClipboard(normalized);
-            return value;
-          });
-        }
-        if (normalized) sendToClipboard(normalized);
-        return result;
-      };
-      try {
-        Object.defineProperty(clipboardProto, 'writeText', {
-          configurable: true,
-          writable: true,
-          value: wrappedProtoWriteText
-        });
-      } catch (_) {
-        clipboardProto.writeText = wrappedProtoWriteText;
-      }
-    } else if (clipboardObj && typeof clipboardObj.writeText === 'function' && !window.__dumberClipboardWriteTextPatched) {
-      window.__dumberClipboardWriteTextPatched = true;
-      var originalWriteText = clipboardObj.writeText.bind(clipboardObj);
-      var wrappedWriteText = function(text) {
-        var normalized = text == null ? '' : String(text);
-        var result = originalWriteText(normalized);
-        if (result && typeof result.then === 'function') {
-          return result.then(function(value) {
-            if (normalized) sendToClipboard(normalized);
-            return value;
-          });
-        }
-        if (normalized) sendToClipboard(normalized);
-        return result;
-      };
-      try {
-        Object.defineProperty(clipboardObj, 'writeText', {
-          configurable: true,
-          writable: true,
-          value: wrappedWriteText
-        });
-      } catch (_) {
-        clipboardObj.writeText = wrappedWriteText;
-      }
-    }
-
-    if (clipboardProto && typeof clipboardProto.write === 'function' && !clipboardProto.__dumberWritePatched) {
-      clipboardProto.__dumberWritePatched = true;
-      var originalProtoWrite = clipboardProto.write;
-      var wrappedProtoWrite = function(items) {
-        var result = originalProtoWrite.apply(this, arguments);
-        if (result && typeof result.then === 'function') {
-          return result.then(function(value) {
-            mirrorClipboardItems(items);
-            return value;
-          });
-        }
-        mirrorClipboardItems(items);
-        return result;
-      };
-      try {
-        Object.defineProperty(clipboardProto, 'write', {
-          configurable: true,
-          writable: true,
-          value: wrappedProtoWrite
-        });
-      } catch (_) {
-        clipboardProto.write = wrappedProtoWrite;
-      }
-    } else if (clipboardObj && typeof clipboardObj.write === 'function' && !window.__dumberClipboardWritePatched) {
-      window.__dumberClipboardWritePatched = true;
-      var originalWrite = clipboardObj.write.bind(clipboardObj);
-      var wrappedWrite = function(items) {
-        var result = originalWrite(items);
-        if (result && typeof result.then === 'function') {
-          return result.then(function(value) {
-            mirrorClipboardItems(items);
-            return value;
-          });
-        }
-        mirrorClipboardItems(items);
-        return result;
-      };
-      try {
-        Object.defineProperty(clipboardObj, 'write', {
-          configurable: true,
-          writable: true,
-          value: wrappedWrite
-        });
-      } catch (_) {
-        clipboardObj.write = wrappedWrite;
-      }
-    }
-  } catch (_) {}
-
-  if (isEditable(document.activeElement)) sendFocusSync();
-})();`
-
-const popupFetchBridgeJS = `(function() {
-  if (window.__dumberPopupBridgePatched) return;
-  window.__dumberPopupBridgePatched = true;
-
-  var bridgeNonce = '__DUMBER_BRIDGE_NONCE__';
-
-  function encodeBody(body) {
-    try {
-      return btoa(unescape(encodeURIComponent(body)));
-    } catch (_) {
-      return '';
-    }
-  }
-
-  function postBridge(url, payload) {
-    var encoded = encodeBody(JSON.stringify(payload));
-    if (!encoded) return;
-    fetch(url, {
-      method: 'POST',
-      headers: {
-        'X-Dumber-Body': encoded,
-        'X-Dumber-Bridge-Nonce': bridgeNonce
-      }
-    }).catch(function() {});
-  }
-
-  function resolvePopupURL(rawURL) {
-    if (rawURL == null || rawURL === '') return '';
-    try {
-      return new URL(String(rawURL), (document && document.baseURI) || location.href).href;
-    } catch (_) {
-      return String(rawURL);
-    }
-  }
-
-  function hasUserGesture() {
-    try {
-      return !!(navigator && navigator.userActivation && navigator.userActivation.isActive);
-    } catch (_) {
-      return false;
-    }
-  }
-
-  function popupHasNoOpener(features) {
-    if (features == null || features === '') return false;
-    var normalized = String(features).toLowerCase();
-    return normalized.indexOf('noopener') !== -1 || normalized.indexOf('noreferrer') !== -1;
-  }
-
-  function shouldDelegateWindowOpen(target) {
-    return target === '_self' || target === '_top' || target === '_parent';
-  }
-
-  function createSyntheticPopupProxy(proxyID, initialURL, features) {
-    var href = initialURL || 'about:blank';
-    var closed = false;
-    var noJavaScriptAccess = popupHasNoOpener(features);
-
-    function navigate(nextURL) {
-      href = resolvePopupURL(nextURL);
-      postBridge('dumb:///api/popup-navigate', { proxy_id: proxyID, url: href });
-      return href;
-    }
-
-    var locationProxy = {
-      assign: function(nextURL) { return navigate(nextURL); },
-      replace: function(nextURL) { return navigate(nextURL); },
-      toString: function() { return href; }
-    };
-    try {
-      Object.defineProperty(locationProxy, 'href', {
-        configurable: true,
-        enumerable: true,
-        get: function() { return href; },
-        set: function(nextURL) { navigate(nextURL); }
-      });
-    } catch (_) {
-      locationProxy.href = href;
-    }
-
-    var proxy = {
-      blur: function() { return undefined; },
-      close: function() {
-        if (closed) return;
-        closed = true;
-        postBridge('dumb:///api/popup-close', { proxy_id: proxyID });
-      },
-      focus: function() { return undefined; },
-      postMessage: function() { return undefined; }
-    };
-    proxy.opener = noJavaScriptAccess ? null : window;
-    proxy.self = proxy;
-    proxy.window = proxy;
-
-    try {
-      Object.defineProperty(proxy, 'closed', {
-        configurable: true,
-        enumerable: true,
-        get: function() { return closed; },
-        set: function(next) { closed = !!next; }
-      });
-    } catch (_) {
-      proxy.closed = false;
-    }
-
-    try {
-      Object.defineProperty(proxy, 'location', {
-        configurable: true,
-        enumerable: true,
-        get: function() { return locationProxy; },
-        set: function(nextURL) { navigate(nextURL); }
-      });
-    } catch (_) {
-      proxy.location = locationProxy;
-    }
-
-    try {
-      Object.defineProperty(proxy, Symbol.toStringTag, {
-        configurable: true,
-        value: 'Window'
-      });
-    } catch (_) {}
-
-    return proxy;
-  }
-
-  try {
-    var originalWindowOpen = typeof window.open === 'function' ? window.open.bind(window) : null;
-    window.open = function(url, target, features) {
-      var normalizedTarget = target == null ? '' : String(target);
-      if (originalWindowOpen && shouldDelegateWindowOpen(normalizedTarget)) {
-        return originalWindowOpen.apply(window, arguments);
-      }
-
-      var proxyID = 'popup-' + Date.now() + '-' + Math.random().toString(36).slice(2);
-      var resolvedURL = resolvePopupURL(url);
-      var noJavaScriptAccess = popupHasNoOpener(features);
-      var popupProxy = createSyntheticPopupProxy(proxyID, resolvedURL, features);
-      postBridge('dumb:///api/popup-open', {
-        proxy_id: proxyID,
-        url: resolvedURL,
-        frame_name: normalizedTarget,
-        user_gesture: hasUserGesture(),
-        no_javascript_access: noJavaScriptAccess
-      });
-      return popupProxy;
-    };
-  } catch (_) {
-    window.__dumberPopupBridgePatched = false;
-  }
-})();`
-
-func buildClipboardSelectionFetchBridgeJS(bridgeNonce string) string {
-	return strings.ReplaceAll(clipboardSelectionFetchBridgeJS, "__DUMBER_BRIDGE_NONCE__", bridgeNonce)
-}
-
-func buildPopupFetchBridgeJS(bridgeNonce string) string {
-	return strings.ReplaceAll(popupFetchBridgeJS, "__DUMBER_BRIDGE_NONCE__", bridgeNonce)
-}
-
 // contentInjector implements port.ContentInjector for the CEF engine.
 // It stores CSS strings and injects them into webviews via ExecuteJavaScript.
 // Thread-safe: InjectThemeCSS may be called from the UI thread while OnLoadEnd
@@ -907,8 +515,7 @@ func (ci *contentInjector) onLoadEnd(wv *WebView) {
 		}
 		logging.FromContext(ctx).Warn().Msg("cef: skipped clipboard bridge injection — nonce generation failed")
 	} else {
-		wv.RunJavaScript(context.Background(), buildClipboardSelectionFetchBridgeJS(bridgeNonce))
-		wv.RunJavaScript(context.Background(), buildPopupFetchBridgeJS(bridgeNonce))
+		wv.RunJavaScript(context.Background(), buildTrustedPageFetchBridgeJS(bridgeNonce))
 	}
 
 	// Video playback diagnostic — logs video element state changes.
