+++ internal/infrastructure/cef/renderer_bridge.go
@@ -26,325 +26,6 @@ const (
 
 var newRendererBridgeProcessMessage = purecef.ProcessMessageCreate
 
-const rendererBridgeExtensionJS = `
-(function() {
-  native function Dispatch(action, payload);
-
-  function send(action, payload) {
-    return Dispatch(action, payload == null ? '' : String(payload));
-  }
-
-  function isEditable(node) {
-    if (!node || node.nodeType !== 1) return false;
-    if (node.isContentEditable) return true;
-    var tag = node.tagName;
-    if (tag !== 'INPUT' && tag !== 'TEXTAREA') return false;
-    if (node.disabled || node.readOnly) return false;
-    return true;
-  }
-
-  function getActiveElementSelection() {
-    var el = document.activeElement;
-    if (!el) return '';
-    var tag = el.tagName;
-    if ((tag === 'INPUT' || tag === 'TEXTAREA') && typeof el.value === 'string') {
-      var start = typeof el.selectionStart === 'number' ? el.selectionStart : 0;
-      var end = typeof el.selectionEnd === 'number' ? el.selectionEnd : 0;
-      if (end > start) return el.value.slice(start, end);
-    }
-    return '';
-  }
-
-  function getSelectedText() {
-    var activeSelection = getActiveElementSelection();
-    if (activeSelection) return activeSelection;
-    var sel = window.getSelection();
-    return sel ? sel.toString() : '';
-  }
-
-  function notifyExplicitTextCopy(action, text) {
-    var normalizedText = text == null ? '' : String(text);
-    var normalizedAction = action == null || action === '' ? 'copy' : String(action);
-    return send('explicit_text_copy', JSON.stringify({ text: normalizedText, action: normalizedAction }));
-  }
-
-  function resolvePopupURL(rawURL) {
-    if (rawURL == null || rawURL === '') return '';
-    try {
-      return new URL(String(rawURL), (document && document.baseURI) || location.href).href;
-    } catch (_) {
-      return String(rawURL);
-    }
-  }
-
-  function hasUserGesture() {
-    try {
-      return !!(navigator && navigator.userActivation && navigator.userActivation.isActive);
-    } catch (_) {
-      return false;
-    }
-  }
-
-  function popupHasNoOpener(features) {
-    if (features == null || features === '') return false;
-    var normalized = String(features).toLowerCase();
-    return normalized.indexOf('noopener') !== -1 || normalized.indexOf('noreferrer') !== -1;
-  }
-
-  function shouldDelegateWindowOpen(target) {
-    return target === '_self' || target === '_top' || target === '_parent';
-  }
-
-  function createSyntheticPopupProxy(proxyID, initialURL, features) {
-    var href = initialURL || 'about:blank';
-    var closed = false;
-    var noJavaScriptAccess = popupHasNoOpener(features);
-
-    function navigate(nextURL) {
-      href = resolvePopupURL(nextURL);
-      send('popup_navigate', JSON.stringify({ proxy_id: proxyID, url: href }));
-      return href;
-    }
-
-    var locationProxy = {
-      assign: function(nextURL) { return navigate(nextURL); },
-      replace: function(nextURL) { return navigate(nextURL); },
-      toString: function() { return href; }
-    };
-    try {
-      Object.defineProperty(locationProxy, 'href', {
-        configurable: true,
-        enumerable: true,
-        get: function() { return href; },
-        set: function(nextURL) { navigate(nextURL); }
-      });
-    } catch (_) {
-      locationProxy.href = href;
-    }
-
-    var proxy = {
-      blur: function() { return undefined; },
-      close: function() {
-        if (closed) return;
-        closed = true;
-        send('popup_close', JSON.stringify({ proxy_id: proxyID }));
-      },
-      focus: function() { return undefined; },
-      postMessage: function() { return undefined; }
-    };
-    proxy.opener = popupHasNoOpener(features) ? null : window;
-    proxy.self = proxy;
-    proxy.window = proxy;
-
-    try {
-      Object.defineProperty(proxy, 'closed', {
-        configurable: true,
-        enumerable: true,
-        get: function() { return closed; },
-        set: function(next) { closed = !!next; }
-      });
-    } catch (_) {
-      proxy.closed = false;
-    }
-
-    try {
-      Object.defineProperty(proxy, 'location', {
-        configurable: true,
-        enumerable: true,
-        get: function() { return locationProxy; },
-        set: function(nextURL) { navigate(nextURL); }
-      });
-    } catch (_) {
-      proxy.location = locationProxy;
-    }
-
-    try {
-      Object.defineProperty(proxy, Symbol.toStringTag, {
-        configurable: true,
-        value: 'Window'
-      });
-    } catch (_) {}
-
-    return proxy;
-  }
-
-  if (typeof window !== 'undefined' && typeof window.open === 'function' && !window.__dumberPopupOpenPatched) {
-    window.__dumberPopupOpenPatched = true;
-    var originalWindowOpen = window.open;
-    window.open = function(url, target, features) {
-      var normalizedTarget = target == null ? '' : String(target);
-      if (shouldDelegateWindowOpen(normalizedTarget)) {
-        return originalWindowOpen.apply(this, arguments);
-      }
-
-      var proxyID = 'popup-' + Date.now() + '-' + Math.random().toString(36).slice(2);
-      var resolvedURL = resolvePopupURL(url);
-      var noJavaScriptAccess = popupHasNoOpener(features);
-      var popupProxy = createSyntheticPopupProxy(proxyID, resolvedURL, features);
-      send('popup_open', JSON.stringify({
-        proxy_id: proxyID,
-        url: resolvedURL,
-        frame_name: normalizedTarget,
-        user_gesture: hasUserGesture(),
-        no_javascript_access: noJavaScriptAccess
-      }));
-      return popupProxy;
-    };
-  }
-
-  if (typeof document !== 'undefined' && document.addEventListener) {
-    document.addEventListener('focusin', function(event) {
-      if (isEditable(event.target)) {
-        send('focus_sync', '');
-      }
-    }, true);
-
-    function mirrorClipboardEvent(action, e) {
-      if (!e.isTrusted) return;
-      var capturedText = '';
-      if (e.clipboardData && typeof e.clipboardData.getData === 'function') {
-        try { capturedText = e.clipboardData.getData('text/plain'); } catch(_) {}
-      }
-      setTimeout(function() {
-        var text = capturedText;
-        if (!text && !e.defaultPrevented) text = getSelectedText();
-        if (text) notifyExplicitTextCopy(action, text);
-      }, 0);
-    }
-
-    document.addEventListener('copy', function(e) {
-      mirrorClipboardEvent('copy', e);
-    }, true);
-    document.addEventListener('cut', function(e) {
-      mirrorClipboardEvent('cut', e);
-    }, true);
-  }
-  try {
-    var clipboardObj = navigator && navigator.clipboard ? navigator.clipboard : null;
-    var clipboardProto = (typeof Clipboard !== 'undefined' && Clipboard.prototype) ||
-      (clipboardObj && typeof Object.getPrototypeOf === 'function' ? Object.getPrototypeOf(clipboardObj) : null);
-
-    function mirrorClipboardItems(items) {
-      try {
-        if (!items || typeof items.length !== 'number') return;
-        Array.prototype.forEach.call(items, function(item) {
-          if (!item || !item.types || item.types.indexOf('text/plain') === -1 || typeof item.getType !== 'function') return;
-          Promise.resolve(item.getType('text/plain'))
-            .then(function(blob) {
-              if (!blob || typeof blob.text !== 'function') return '';
-              return blob.text();
-            })
-            .then(function(text) {
-              if (text) notifyExplicitTextCopy('copy', text);
-            })
-            .catch(function() {});
-        });
-      } catch (_) {}
-    }
-
-    if (clipboardProto && typeof clipboardProto.writeText === 'function' && !clipboardProto.__dumberWriteTextPatched) {
-      clipboardProto.__dumberWriteTextPatched = true;
-      var originalProtoWriteText = clipboardProto.writeText;
-      var wrappedProtoWriteText = function(text) {
-        var normalized = text == null ? '' : String(text);
-        var result = originalProtoWriteText.call(this, normalized);
-        if (result && typeof result.then === 'function') {
-          return result.then(function(value) {
-            if (normalized) notifyExplicitTextCopy('copy', normalized);
-            return value;
-          });
-        }
-        if (normalized) notifyExplicitTextCopy('copy', normalized);
-        return result;
-      };
-      try {
-        Object.defineProperty(clipboardProto, 'writeText', {
-          configurable: true,
-          writable: true,
-          value: wrappedProtoWriteText
-        });
-      } catch (_) {
-        clipboardProto.writeText = wrappedProtoWriteText;
-      }
-    } else if (clipboardObj && typeof clipboardObj.writeText === 'function' && !window.__dumberClipboardWriteTextPatched) {
-      window.__dumberClipboardWriteTextPatched = true;
-      var originalWriteText = clipboardObj.writeText.bind(clipboardObj);
-      var wrappedWriteText = function(text) {
-        var normalized = text == null ? '' : String(text);
-        var result = originalWriteText(normalized);
-        if (result && typeof result.then === 'function') {
-          return result.then(function(value) {
-            if (normalized) notifyExplicitTextCopy('copy', normalized);
-            return value;
-          });
-        }
-        if (normalized) notifyExplicitTextCopy('copy', normalized);
-        return result;
-      };
-      try {
-        Object.defineProperty(clipboardObj, 'writeText', {
-          configurable: true,
-          writable: true,
-          value: wrappedWriteText
-        });
-      } catch (_) {
-        clipboardObj.writeText = wrappedWriteText;
-      }
-    }
-
-    if (clipboardProto && typeof clipboardProto.write === 'function' && !clipboardProto.__dumberWritePatched) {
-      clipboardProto.__dumberWritePatched = true;
-      var originalProtoWrite = clipboardProto.write;
-      var wrappedProtoWrite = function(items) {
-        var result = originalProtoWrite.apply(this, arguments);
-        if (result && typeof result.then === 'function') {
-          return result.then(function(value) {
-            mirrorClipboardItems(items);
-            return value;
-          });
-        }
-        mirrorClipboardItems(items);
-        return result;
-      };
-      try {
-        Object.defineProperty(clipboardProto, 'write', {
-          configurable: true,
-          writable: true,
-          value: wrappedProtoWrite
-        });
-      } catch (_) {
-        clipboardProto.write = wrappedProtoWrite;
-      }
-    } else if (clipboardObj && typeof clipboardObj.write === 'function' && !window.__dumberClipboardWritePatched) {
-      window.__dumberClipboardWritePatched = true;
-      var originalWrite = clipboardObj.write.bind(clipboardObj);
-      var wrappedWrite = function(items) {
-        var result = originalWrite(items);
-        if (result && typeof result.then === 'function') {
-          return result.then(function(value) {
-            mirrorClipboardItems(items);
-            return value;
-          });
-        }
-        mirrorClipboardItems(items);
-        return result;
-      };
-      try {
-        Object.defineProperty(clipboardObj, 'write', {
-          configurable: true,
-          writable: true,
-          value: wrappedWrite
-        });
-      } catch (_) {
-        clipboardObj.write = wrappedWrite;
-      }
-    }
-  } catch (_) {}
-
-  send('bridge_ready', (typeof location !== 'undefined' && location && location.href) ? location.href : '');
-})();
-`
-
 type rendererBridgeProcessHandler struct {
 	registerOnce  sync.Once
 	bridgeHandler purecef.V8Handler
